pyproject.toml
[tool.poetry]
name = "hack2"
version = "0.1.0"
description = ""
authors = ["Tim <jlemyp96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
djangorestframework_simplejwt = "^4.3"
django-reversion = "^3.0"
django-filter = "^2.2"
django-extensions = "^2.2"
django-suit = { git = "https://github.com/darklow/django-suit", branch = "v2" }

[tool.poetry.dev-dependencies]
Werkzeug = "^0.15.5"
ipython = "^7.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
